+++ drawio-custom-plugins/src/index.ts
@@ -2,3 +2,4 @@ import "./linkSelectedNodeWithData";
 import "./liveshare";
 import "./focus";
 import "./menu-entries";
+import "./donationDialog";