+++ src/index.ts
@@ -2,15 +2,13 @@ import * as vscode from "vscode";
 import { MobxConsoleLogger } from "@knuddels/mobx-logger";
 import * as mobx from "mobx";
 import { Extension } from "./Extension";
-import { join } from "path";
 
 if (process.env.DEV === "1") {
 	new MobxConsoleLogger(mobx);
 }
 
 export function activate(context: vscode.ExtensionContext) {
-	const packageJsonPath = join(context.extensionPath, "package.json");
-	context.subscriptions.push(new Extension(packageJsonPath));
+	context.subscriptions.push(new Extension(context));
 }
 
 export function deactivate() {}